requirements.txt
pyserial
dependencies/pyserial-3.5-py2.py3-none-any.whl
matplotlib